episode 82

+++ database/migrations/2020_06_24_135124_create_thread_subscriptions_table.php
@@ -16,8 +16,8 @@ class CreateThreadSubscriptionsTable extends Migration
         // Schema::dropIfExists('thread_subscriptions');
         Schema::create('thread_subscriptions', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('thread_id');
+            $table->unsignedInteger('user_id');
+            $table->unsignedInteger('thread_id');
             $table->timestamps();
             $table->unique(['user_id', 'thread_id']);
